Fixes #4: modularize tests
